python_app/requirements.txt
Flask==1.1.2
pdfkit==0.6.1
gunicorn==20.0.4
gevent==20.4.0
requests==2.25.0
lxml==4.5.0
python-slugify==4.0.1
